Use a list instead of a vector to identify fixed-effects in cpp_demean
This is an important update. It saves one deep copy of the data without performance cost. So it's a direct improvement.

I modified both the C++ code and the R code calling it.

I have been careful in checking everything was properly modified in the C++ code, but as usual... we never know.